+++ scandeval/benchmark.py
@@ -113,14 +113,19 @@ class Benchmark:
         Returns:
             list of str: The model IDs of the relevant models.
         '''
-        url = 'https://huggingface.co/models'
+        # Set GET request parameter values
         params = dict()
         if language is not None:
             params['filter'] = language
         if task is not None:
             params['pipeline_tag'] = task
+
+        # Fetch and parse the html from the HuggingFace Hub
+        url = 'https://huggingface.co/models'
         html = requests.get(url, params=params).text
         soup = BeautifulSoup(html, 'html.parser')
+
+        # Extract the model ids from the html
         articles = soup.find_all('article')
         model_ids = [header['title']
                      for article in articles
@@ -128,6 +133,22 @@ class Benchmark:
                      if header.get('class') is not None and
                      header.get('title') is not None and
                      'items-center' in header['class']]
+
+        # There might still be models in the list, which is *not* in the given
+        # language, but still ended up in the search results, as the
+        # HuggingFace Hub only searches generally for the language code. We
+        # filter these out here.
+        for model_id in model_ids:
+            model_url = f'https://huggingface.co/{model_id}'
+            html = requests.get(model_url).text
+            soup = BeautifulSoup(html, 'html.parser')
+            a_tags_with_class = [a for a in soup.find_all('a')
+                                 if a.get('class') is not None]
+            languages = [a['tag-id'] for a in a_tags_with_class
+                         if 'tag-green' in a['class']]
+            if language not in languages:
+                model_ids.remove(model_id)
+
         return model_ids
 
     def _get_model_lists(self,